Move encoding functions to ovsp4rt_encoders.cc
- Moved the protobuf encoding functions from `newovsp4rt.cc` to
  `ovsp4rt_encoders.h`.

- Used Copilot to verify that the extracted functions are
  identical to their counterparts in `ovsp4rt.cc`.
